requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
idna==3.10
logzero==1.7.0
numpy==2.2.3
pandas==2.2.3
pip==24.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
smartapi-python==1.5.2
tzdata==2025.1
urllib3==2.3.0
